replace start with first and second
Fibonacci sequences depend on the first and second number in the
sequence. For example, it can start at 1 and 1 or 1 and 2; therefore,
the starting number in the sequence is too ambiguous to depend on.

This commit both changes the API and reorders proto numbers. This
is breaking but is being treated as if the API has not yet been
released. That's realistic because the API hasn't been constructed
yet (no starting gRPC server) and there is not yet any consumer
implementation. All packages are thus far internal and cannot be
used as libraries.